refactor: make python module for api
